File/Convert/Solution.php

+++ File/Convert/Solution.php
@@ -132,7 +132,9 @@ class File_Convert_Solution
    
     static function deleteOnExit()
     {
-        var_dump(self::$deleteOnExit);
+        if (empty(self::$deleteOnExit)) {
+            return; // should not happend but appears to?
+        }
         foreach(self::$deleteOnExit as $fn) {
             if (file_exists($fn)) {
                 unlink($fn);
